--- report-service/src/main/java/gov/cdc/dataingestion/authintegration/controller/AuthController.java
@@ -1,40 +1,0 @@
-package gov.cdc.dataingestion.authintegration.controller;
-
-import gov.cdc.dataingestion.authintegration.service.AuthService;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.config.Registry;
-import org.apache.http.config.RegistryBuilder;
-import org.apache.http.conn.socket.ConnectionSocketFactory;
-import org.apache.http.conn.socket.PlainConnectionSocketFactory;
-import org.apache.http.conn.ssl.NoopHostnameVerifier;
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
-import org.apache.http.conn.ssl.TrustStrategy;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.impl.conn.BasicHttpClientConnectionManager;
-import org.apache.http.ssl.SSLContexts;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.net.ssl.SSLContext;
-
-@RestController
-public class AuthController {
-    private static Logger logger = LoggerFactory.getLogger(AuthController.class);
-
-    private final AuthService authService;
-
-    @Autowired
-    public AuthController(AuthService authService) {
-        this.authService = authService;
-    }
-
-    @GetMapping("/auth/token")
-    public String getAuthToken() {
-        logger.info("Inside getAuthToken() controller...");
-        return authService.getToken();
-    }
-}
